airflow_scripts/requirements.txt
numpy
pandas
scikit-learn==1.5.1
matplotlib
seaborn
price-predictions
category-encoders
dvc
dvc-gdrive
apache-airflow-providers-google